refactor: split extract/lsp.rs into lsp/mod.rs and lsp/transport.rs
Move JSON-RPC transport layer (LspTransport, PipelinedTransport, URI
helpers) into src/extract/lsp/transport.rs. The enricher, state, all
five passes, and tests remain in src/extract/lsp/mod.rs.

The transport module is self-contained: sequential init-phase framing,
concurrent pipelined framing, quiescent_flag live-updating, diagnostics
sink, and URI helpers (path_to_uri, find_enclosing_symbol,
uri_to_relative_path).

No behavior changes. Closes #411.

[outcome:context-assembly]
